+++ CHANGELOG.md
@@ -20,10 +20,14 @@ However, given the rather burst-like development on this package, features are o
 - !366 makes it possible to pass arbitrary keyword arguments to plot functions, even if the same keys are used by dantro (e.g. the `style` key). To do so, keyword arguments defined in `plot_func_kwargs` are now directly passed on to the plot function.
 - !366 adds the `max_num_nodes` argument to DAG visualization; if the number of nodes of the DAG exceeds this, the visualization will not be performed.
 - !366 adds the `.to_datarray` and `get_operation` operations to the data transformation framework.
+- !368 adds the `snsplot` plot function, bridging to [seaborn](https://seaborn.pydata.org); it is available via the `.plot.snsplot` base config.
 
 
 #### Bug fixes
 - !367 fixes `is_iterable` after a [behaviour change in Python 3.13.4](https://github.com/python/cpython/issues/135171).
+- !368 fixes a bug where non-pickleable objects (e.g. lambdas) were retained in `Transformation` objects' memory cache, hence prohibiting multiprocessing for parallel plotting.
+    - To resolve this, `Transformation` now excludes these objects from both the memory and file cache and instead re-evaluates them each time in order to not retain them.
+    As they are fast to construct, the performance drawback is negligible.
 - !369 resolves an issue in `.plot.errorbars` where labels were (undesirably) ordered alphabetically along the `hue` dimension. Thanks to [@tgaskin](https://gitlab.com/tgaskin)!